paging: switch map_zone arguments to 64 bit, to avoid overflow

+++ kernel/paging.c
@@ -84,8 +84,9 @@ void map_page(uint32_t *dir, uint32_t virt_addr, uint32_t phy_addr,
     page_table[pt_index] = phy_addr | flags | PRESENT;
 }
 
-void map_zone(uint32_t *pdir, uint32_t virt_start, uint32_t virt_end,
-              uint32_t phy_start, uint32_t phy_end, int align_virt_end, uint32_t flags)
+void map_zone(uint32_t *pdir, uint64_t virt_start, uint64_t virt_end,
+              uint64_t phy_start, uint64_t phy_end, int align_virt_end,
+              uint32_t flags)
 {
     if (virt_end - virt_start != phy_end - phy_start) {
         panic("map_zone physical and virtual zones must be the same size!");
@@ -98,7 +99,7 @@ void map_zone(uint32_t *pdir, uint32_t virt_start, uint32_t virt_end,
     if (align_virt_end)
         virt_end   = ALIGN_UP(virt_end) - 1;
 
-    for (uint32_t virt = virt_start, phy = phy_start; virt <= virt_end;
+    for (uint64_t virt = virt_start, phy = phy_start; virt <= virt_end;
          virt += PAGE_SIZE, phy += PAGE_SIZE) {
         map_page(pdir, virt, phy, flags);
     }
